requirements.txt
pandas
geopandas
shapely
contextily
matplotlib
docxtpl
docx
python-dotenv
gevent
socketIO_client